TODOList/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
install==1.3.5
pytz==2021.3
sqlparse==0.4.2
